requirements.txt
boto==2.48.0
boto3==1.6.21
botocore==1.9.21
Collectfast==0.6.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.3
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-ckeditor==5.4.0
django-imagekit==4.0.2
django-js-asset==1.0.0
django-leaflet==0.23.0
django-redis-cache==1.7.1
django-storages==1.6.6
django-toolbelt==0.0.1
django-widget-tweaks==1.4.1
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylibmc==1.5.2
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2018.3
redis==2.10.6
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
uWSGI==2.0.17
whitenoise==3.3.1
